Correct_Check_Hit

+++ popcorn_dos_game/Platform.cpp
@@ -15,7 +15,7 @@ bool AsPlatform::Check_Hit(double next_x_pos, double next_y_pos, ABall *ball)
 // Correction position when reflecting from the platform
 	if (next_y_pos + ball -> Radius > AsConfig::Platform_Y_Pos)
 	{
-		if (next_x_pos + ball -> Radius >= X_Pos && next_x_pos - ball -> Radius<= X_Pos + Width)
+		if (next_x_pos + ball -> Radius >= X_Pos && next_x_pos - ball -> Radius<= (double)(X_Pos + Width) )
 		{
 			ball -> Reflect(true);
 			return true;
